job/requirements.txt
google-cloud-bigquery>=2.0.0
aiohttp>=3.8.0
pytz>=2021.1